+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project <em>does not yet</em> adheres to [Semantic Versioning](https://semv
 
 ### Added
 
+- `/api/v1/dwc_occurrences/area_autocomplete?term=Can&target=country|stateProvince|county`
 - Radial filter: Source => TNR and vice versa
 - Taxonomic tree: Options to show/hide descendants count and Level-based coloring.
 - Interactive key descriptor selections can be used to help create a dichotomous key